+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [12.3.6] - 2018-??
+### Added
+- Audit.NET.DynamoDB: Adding new extension **Audit.NET.DynamoDB** to save audit events into Amazon DynamoDB using the `AWSSDK.DynamoDBv2` document model.
+
 ### Modified
 - Audit.Core: Adding GetXxxxxEvent() extensions to AuditScope in addition to the AuditEvent.
 - Audit.EntityFramework: Added validation on EF Data Provider to ignore non EF events.